refactor: isolate Table stories
