+++ packages/react-aria-components/stories/utils.tsx
@@ -1,5 +1,5 @@
 import {classNames} from '@react-spectrum/utils';
-import {ListBoxItem, ListBoxItemProps} from 'react-aria-components';
+import {ListBoxItem, ListBoxItemProps, MenuItem, MenuItemProps} from 'react-aria-components';
 import React from 'react';
 import styles from '../example/index.css';
 
@@ -11,6 +11,17 @@ export const MyListBoxItem = (props: ListBoxItemProps) => {
         focused: isFocused,
         selected: isSelected,
         hovered: isHovered
+      })} />
+  );
+};
+
+export const MyMenuItem = (props: MenuItemProps) => {
+  return (
+    <MenuItem
+      {...props}
+      className={({isFocused, isSelected}) => classNames(styles, 'item', {
+        focused: isFocused,
+        selected: isSelected
       })} />
   );
 };
